+++ App.js
@@ -21,6 +21,7 @@ import { AppNavigator } from './navigation/Navigator';
 import RNFS from 'react-native-fs';
 import SplashScreen from 'react-native-splash-screen';
 import Orientation from 'react-native-orientation-locker';
+import { RNFFmpegConfig } from 'react-native-ffmpeg';
 
 /**
  * Use any valid `name` property from eva icons (e.g `github`, or `heart-outline`)
@@ -76,6 +77,7 @@ export default () => {
 
   useEffect(() => {
     AppState.addEventListener('change', handleAppState);
+    RNFFmpegConfig.disableLogs();
     if (!Platform.isPad) Orientation.lockToPortrait();
     SplashScreen.hide();
     return () => {
