+++ components/postHorizontalList.js
@@ -4,8 +4,8 @@ import { Layout, Text, Button, Icon } from '@ui-kitten/components';
 import { SmallCard } from './cardHorizontal';
 import { useNavigation } from '@react-navigation/native';
 import { getPosts } from '../api/post';
-import { tagStyles } from '../styles';
 import { getData } from '../util/storage';
+import { postWithDetails } from '../util/post';
 
 const PlusIcon = (props) => <Icon {...props} name="plus-circle-outline" />;
 const UmbrellaIcon = (props) => <Icon {...props} name="umbrella-outline" />;
@@ -29,38 +29,6 @@ export const PostHorizontalList = forwardRef((props, ref) => {
     },
   }));
 
-  const postWithDetails = (tagsWithType, post, votes) => {
-    var artist = '';
-    var copyright = '';
-    var tags = [];
-    const postTags = post.tags.split(' ');
-    for (const tag of postTags) {
-      const type = tagsWithType[tag];
-      var style = tagStyles.artist_outline;
-      if (type === 'artist') artist = artist + ' ' + capitalize(tag);
-      if (type === 'copyright') {
-        style = tagStyles.copyright_outline;
-        copyright = tag;
-      }
-      if (type === 'terminology') style = tagStyles.terminology_outline;
-      if (type === 'meta') style = tagStyles.meta_outline;
-      if (type === 'general') style = tagStyles.general_outline;
-      tags.push({ type, tag, style });
-    }
-
-    const name =
-      artist.trim() && artist.trim() !== 'Artist_unknown'
-        ? artist.replace('Artist_unknown', '').trim()
-        : copyright.trim();
-    const title = name ? capitalize(name).replace(/_/g, ' ') : name;
-
-    tags.sort((a, b) => a.type > b.type);
-    post.userScore = votes[post.id] ? votes[post.id] : 0;
-    post.tags = tags;
-    post.title = title;
-    return post;
-  };
-
   const fetchPost = async (page, isFirst, search) => {
     try {
       if (!isFirst) setFetching(true);
@@ -72,10 +40,18 @@ export const PostHorizontalList = forwardRef((props, ref) => {
       const user = await getData('user');
       if (user) params = { ...params, user: user.name, password_hash: user.password_hash };
       const response = await getPosts(params);
-      const postsWithTitle = response.posts.map((p) => postWithDetails(response.tags, p, response.votes));
-      const filteredPosts = postsWithTitle.filter((p) => !data.some((currentPost) => currentPost.id === p.id));
-      let newData = [...data, ...filteredPosts];
-      if (page === 1) newData = postsWithTitle;
+      var newData = [];
+      if (page === 1) {
+        newData = response.posts.map((p) => postWithDetails(response.tags, p, response.votes));
+      } else {
+        var newPosts = [];
+        for (const p of response.posts) {
+          const newPost = postWithDetails(response.tags, p, response.votes);
+          const index = data.findIndex((post) => post.id === newPost.id);
+          if (index === -1) newPosts.push(newPost);
+        }
+        newData = data.concat(newPosts);
+      }
       if (!message || message === 'Error, please try again later :(') setMessage('Nobody here but us chickens!');
       setData(newData);
       clearLoading();
