+++ screens/EditPost.js
@@ -29,7 +29,7 @@ import { useNavigation } from '@react-navigation/native';
 import { getRelatedTags, getTags } from '../api/tag';
 import { updatePost } from '../api/post';
 import Toast from 'react-native-simple-toast';
-import { tagStyles } from '../styles';
+import { postWithDetails } from '../util/post';
 import { scale } from 'react-native-size-matters';
 import Orientation from 'react-native-orientation-locker';
 
@@ -80,37 +80,6 @@ export const EditPostScreen = ({ route }) => {
   const navigateLogin = () => {
     setEditLoading(false);
     navigation.navigate('Login', { from: 'EditPost' });
-  };
-
-  const postWithDetails = (tagsWithType, post, votes = []) => {
-    var artist = '';
-    var copyright = '';
-    var tags = [];
-    const postTags = post.tags.split(' ');
-    for (const tag of postTags) {
-      const type = tagsWithType[tag];
-      var style = tagStyles.artist_outline;
-      if (type === 'artist') artist = artist + ' ' + capitalize(tag);
-      if (type === 'copyright') {
-        style = tagStyles.copyright_outline;
-        copyright = tag;
-      }
-      if (type === 'terminology') style = tagStyles.terminology_outline;
-      if (type === 'meta') style = tagStyles.meta_outline;
-      if (type === 'general') style = tagStyles.general_outline;
-      tags.push({ type, tag, style });
-    }
-
-    const name =
-      artist.trim() && artist.trim() !== 'Artist_unknown'
-        ? artist.replace('Artist_unknown', '').trim()
-        : copyright.trim();
-    const title = name ? capitalize(name).replace(/_/g, ' ') : name;
-    tags.sort((a, b) => a.type > b.type);
-    post.userScore = votes[post.id] ? votes[post.id] : 0;
-    post.tags = tags;
-    post.title = title;
-    return post;
   };
 
   const editPostData = async () => {
